requirements.txt
python-dotenv
pydantic
openai-agents
browser-use
langchain-openai
openai
pillow
langchain-google-genai